pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["featuregating"]

[project]
name = "featuregating"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.5.0",
    "einops>=0.8.1",
    "gdown>=5.2.0",
    "huggingface-hub>=0.34.4",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pytest>=8.4.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "timm>=1.0.15",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
    "vit-prisma",
    "wandb>=0.20.1",
]

[tool.uv.sources]
vit-prisma = { git = "https://github.com/piragi/ViT-Prisma.git", branch = "main" }

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for full pipeline"
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
